requirements.txt
beautifulsoup4==4.6.3
bs4==0.0.1
certifi==2018.10.15
chardet==3.0.4
cycler==0.10.0
idna==2.7
jieba==0.39
kiwisolver==1.0.1
matplotlib==3.0.0
numpy==1.15.3
Pillow==5.3.0
pyparsing==2.2.2
python-dateutil==2.7.3
requests==2.20.0
six==1.11.0
urllib3==1.24
wordcloud==1.5.0
